+++ spec/vips_spec.rb
@@ -71,6 +71,15 @@ RSpec.describe Vips do
       expect(embed.bands).to eq(1)
     end
 
+    it 'can handle enum arguments with underscores' do
+      black = Vips::Image.black 200, 100, bands: 3
+      mono = black.colourspace :b_w
+
+      expect(mono.width).to eq(200)
+      expect(mono.height).to eq(100)
+      expect(mono.bands).to eq(1)
+    end
+
     it 'enum arguments can be strings' do
       black = Vips::Operation.call "black", [200, 200]
       embed = Vips::Operation.call "embed", [black, 10, 10, 500, 500],